refactor: moved some types to controllers
