+++ src/types/controllerTypes.ts
@@ -1,64 +1,4 @@
-import type { IProvider } from './coreTypes';
-
 export type CaipAddress = `${string}:${string}:${string}`;
-
-// -- ClientCtrl ------------------------------------------- //
-export interface ClientCtrlState {
-  initialized: boolean;
-  provider?: IProvider;
-  sessionTopic?: string;
-}
-
-// -- ModalCtrl --------------------------------------- //
-export interface ModalCtrlState {
-  open: boolean;
-}
-
-// -- OptionsCtrl --------------------------------------- //
-export interface OptionsCtrlState {
-  isDataLoaded: boolean;
-}
-
-// -- AccountCtrl --------------------------------------- //
-export interface AccountCtrlState {
-  address?: string;
-  isConnected: boolean;
-}
-
-// -- WcConnectionCtrl ------------------------------------- //
-export interface WcConnectionCtrlState {
-  pairingUri: string;
-  pairingEnabled: boolean;
-  pairingError: boolean;
-}
-
-// -- ThemeCtrl --------------------------------------------- //
-export interface ThemeCtrlState {
-  themeMode?: 'dark' | 'light';
-  accentColor?: string;
-}
-
-// -- ToastCtrl ------------------------------------------ //
-export interface ToastCtrlState {
-  open: boolean;
-  message: string;
-  variant: 'error' | 'success';
-}
-
-// -- RouterCtrl --------------------------------------------- //
-export type RouterView =
-  | 'ConnectWallet'
-  | 'Qrcode'
-  | 'WalletExplorer'
-  | 'Connecting';
-
-export interface RouterCtrlState {
-  history: RouterView[];
-  view: RouterView;
-  data?: {
-    wallet?: WcWallet;
-  };
-}
 
 // -- ApiCtrl Types -------------------------------------------------------
 export interface WcWallet {
